+++ CHANGELOG.md
@@ -4,13 +4,17 @@
 All notable changes to this project will be documented in this file.
 <!--- END HEADER -->
 
-## [0.0.2](https://github.com/Faktly-GmbH/laravel-prometheus-metrics/compare/v0.0.1...v0.0.2) (2025-12-29)
+## [0.0.3](https://github.com/Faktly-GmbH/laravel-prometheus-metrics/compare/v0.0.2...v0.0.3) (2025-12-29)
 
+- Add migration to count user sessions even with SESSION_DRIVER set to "cookie"
+- Add command to clean-up tracked user sessions
 
 ---
 
-## [0.0.1](https://github.com/Faktly-GmbH/laravel-prometheus-metrics/compare/0.0.0...v0.0.1) (2025-12-29)
+## [0.0.2](https://github.com/Faktly-GmbH/laravel-prometheus-metrics/compare/v0.0.1...v0.0.2) (2025-12-29)
 
+- Set default output format to "prometheus"
+- Fix implicit nullable parameter type hint
 
 ---